Рефакторинг: Убрал ненужный класс MyCanvas Вынес QSFMLCanvas в отдельный модуль Поправил CMakeLists.txt
